// interfaces/global.ts
export interface Toast {
    ADD_TIME: string;
    TYPE: string;
    FROM: string;
    TEXT: string;
}

export interface ToastSettings {
    ENABLE_TOASTS: boolean;
    TIMEOUT: number;
    SHOW_DATA: {
        [key: string]: boolean;
    };
    POSITION: {
        VERTICAL: string;
        HORIZONTAL: string;
    };
}